+++ crates/lexgen/src/ast.rs
@@ -24,8 +24,6 @@ pub enum Rule<Rhs> {
 
     /// `type Error = UserError;`
     ErrorType {
-        /// Lifetimes on the LHS of the declaration, e.g. `type Error<'input> = ...`;
-        lifetimes: Vec<syn::Lifetime>,
         /// Type on the RHS, e.g. `UserError<'input>`
         ty: syn::Type,
     },
@@ -92,11 +90,7 @@ impl<Rhs: fmt::Debug> fmt::Debug for Rule<Rhs> {
                 .debug_struct("Rule::UnnamedRules")
                 .field("rules", rules)
                 .finish(),
-            Rule::ErrorType { lifetimes, ty } => f
-                .debug_struct("Rule::ErrorType")
-                .field("lifetimes", lifetimes)
-                .field("ty", ty)
-                .finish(),
+            Rule::ErrorType { ty } => f.debug_struct("Rule::ErrorType").field("ty", ty).finish(),
         }
     }
 }
@@ -329,17 +323,10 @@ impl Parse for Rule<RuleRhs> {
             if ident != "Error" {
                 panic!("Error type syntax is: `type Error = ...;`");
             }
-            let mut lifetimes: Vec<syn::Lifetime> = vec![];
-            if input.parse::<syn::token::Lt>().is_ok() {
-                while !input.peek(syn::token::Gt) {
-                    lifetimes.push(input.parse()?);
-                }
-                input.parse::<syn::token::Gt>()?;
-            }
             input.parse::<syn::token::Eq>()?;
             let ty = input.parse::<syn::Type>()?;
             input.parse::<syn::token::Semi>()?;
-            Ok(Rule::ErrorType { ty, lifetimes })
+            Ok(Rule::ErrorType { ty })
         } else {
             let mut single_rules = vec![];
             while !input.is_empty() {