+++ crates/lexgen/src/dfa/codegen/ctx.rs
@@ -21,10 +21,6 @@ pub struct CgCtx {
     /// Type of the user error, when available. `<type>` in `type Error = ...`.
     user_error_type: Option<syn::Type>,
 
-    /// Name of the `enum` type for user actions: `enum LexerAction { ... }`. The name is derived
-    /// from `lexer_name`.
-    action_type_name: syn::Ident,
-
     /// Maps user-written rule names (e.g. `rule MyRule { ... }`) to their initial states in the
     /// final DFA.
     rule_states: FxHashMap<String, StateIdx>,
@@ -56,9 +52,6 @@ impl CgCtx {
         user_error_type: Option<syn::Type>,
         rule_states: FxHashMap<String, StateIdx>,
     ) -> CgCtx {
-        let action_type_name =
-            syn::Ident::new(&(lexer_name.to_string() + "Action"), lexer_name.span());
-
         let inlined_states: Vec<StateIdx> = dfa
             .states
             .iter()
@@ -77,7 +70,6 @@ impl CgCtx {
             lexer_name,
             token_type,
             user_error_type,
-            action_type_name,
             rule_states,
             inlined_states,
             codegen_state: CgState {
@@ -109,10 +101,6 @@ impl CgCtx {
         self.user_error_type.as_ref()
     }
 
-    pub fn action_type_name(&self) -> &syn::Ident {
-        &self.action_type_name
-    }
-
     pub fn add_search_table(&mut self, ranges: Vec<(char, char)>) -> syn::Ident {
         self.codegen_state.search_tables.add_table(ranges)
     }
@@ -127,5 +115,12 @@ impl CgCtx {
 
     pub fn iter_semantic_actions(&self) -> impl Iterator<Item = (SemanticActionIdx, &RuleRhs)> {
         self.semantic_action_table.iter()
+    }
+
+    pub fn semantic_action_fn_ident(&self, action: SemanticActionIdx) -> syn::Ident {
+        syn::Ident::new(
+            &format!("{}_ACTION_{}", self.lexer_name, action.as_usize()),
+            self.lexer_name.span(),
+        )
     }
 }