+++ crates/lexgen/tests/bugs.rs
@@ -1,7 +1,7 @@
 mod test_utils;
 
 use lexgen::lexer;
-use test_utils::next;
+use test_utils::{loc, next};
 
 #[test]
 fn failure_confusion_1() {
@@ -107,13 +107,13 @@ fn failure_confusion_3_1() {
     }
 
     let mut lexer = Lexer::new("a ab abc");
-    assert_eq!(lexer.next(), Some(Ok((0, 2, 1))));
-    assert_eq!(lexer.next(), Some(Ok((1, 0, 2))));
-    assert_eq!(lexer.next(), Some(Ok((2, 1, 4))));
-    assert_eq!(lexer.next(), Some(Ok((4, 0, 5))));
-    assert_eq!(lexer.next(), Some(Ok((5, 1, 7))));
-    assert_eq!(lexer.next(), Some(Ok((7, 2, 8))));
-    assert_eq!(lexer.next(), None);
+    assert_eq!(next(&mut lexer), Some(Ok(2)));
+    assert_eq!(next(&mut lexer), Some(Ok(0)));
+    assert_eq!(next(&mut lexer), Some(Ok(1)));
+    assert_eq!(next(&mut lexer), Some(Ok(0)));
+    assert_eq!(next(&mut lexer), Some(Ok(1)));
+    assert_eq!(next(&mut lexer), Some(Ok(2)));
+    assert_eq!(next(&mut lexer), None);
 }
 
 #[test]
@@ -129,9 +129,9 @@ fn failure_confusion_3_2() {
     }
 
     let mut lexer = Lexer::new("f,");
-    assert_eq!(lexer.next(), Some(Ok((0, 1, 1))));
-    assert_eq!(lexer.next(), Some(Ok((1, 2, 2))));
-    assert_eq!(lexer.next(), None);
+    assert_eq!(next(&mut lexer), Some(Ok(1)));
+    assert_eq!(next(&mut lexer), Some(Ok(2)));
+    assert_eq!(next(&mut lexer), None);
 }
 
 #[test]
@@ -147,10 +147,10 @@ fn failure_confusion_4() {
 
     let mut lexer = Lexer::new("aaa aa a");
 
-    assert_eq!(lexer.next(), Some(Ok((0, 1, 3))));
-    assert_eq!(lexer.next(), Some(Ok((4, 2, 6))));
-    assert_eq!(lexer.next(), Some(Ok((7, 3, 8))));
-    assert_eq!(lexer.next(), None);
+    assert_eq!(next(&mut lexer), Some(Ok(1)));
+    assert_eq!(next(&mut lexer), Some(Ok(2)));
+    assert_eq!(next(&mut lexer), Some(Ok(3)));
+    assert_eq!(next(&mut lexer), None);
 }
 
 #[test]
@@ -227,16 +227,18 @@ fn return_should_reset_match() {
 
 #[test]
 fn issue_16_backtracking_1() {
-    fn return_match<'lexer, 'input>(lexer: &mut Lexer<'input>) -> LexerAction<&'input str> {
-        let match_ = lexer.match_();
-        lexer.return_(match_)
-    }
-
     lexer! {
         Lexer -> &'input str;
 
-        'a'+ 'b' => return_match,
-        'a' => return_match,
+        'a'+ 'b' => |lexer| {
+            let match_ = lexer.match_();
+            lexer.return_(match_)
+        },
+
+        'a' => |lexer| {
+            let match_ = lexer.match_();
+            lexer.return_(match_)
+        },
     }
 
     let mut lexer = Lexer::new("aaaab");
@@ -253,7 +255,9 @@ fn issue_16_backtracking_1() {
 
 #[test]
 fn issue_16_backtracking_2() {
-    fn return_match<'lexer, 'input>(lexer: &mut Lexer<'input>) -> LexerAction<&'input str> {
+    fn return_match<'input>(
+        lexer: &mut Lexer<'input>,
+    ) -> lexgen_util::SemanticActionResult<&'input str> {
         let match_ = lexer.match_();
         lexer.return_(match_)
     }
@@ -295,7 +299,13 @@ fn end_of_input_handling() {
     }
 
     let mut lexer = Lexer::new("aa");
-    assert_eq!(lexer.next(), Some(Ok((0, (0, "a"), 1))));
-    assert_eq!(lexer.next(), Some(Ok((1, (1, "a"), 2))));
+    assert_eq!(
+        lexer.next(),
+        Some(Ok((loc(0, 0, 0), (0, "a"), loc(0, 1, 1))))
+    );
+    assert_eq!(
+        lexer.next(),
+        Some(Ok((loc(0, 1, 1), (1, "a"), loc(0, 2, 2))))
+    );
     assert_eq!(lexer.next(), None);
 }