+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# Unreleased
+
+- Common code in generated code is moved to a new crate `lexgen_util`.
+  lexgen-generated lexers now depend on `lexgen_util`.
+
+- Line and column tracking implemented. Iterator implementation now yields
+  `(Loc, Token, Loc)`, where `Loc` is defined in `lexgen_util` as `struct Loc {
+  line: u32, col: u32, byte_idx: usize }`.
+
 # 2021/10/21: 0.7.0
 
 - Regex syntax updated to include "any character" (`_`) and "end of input"